[feat] 레이싱 게임 구현 (#845)
* chore: 패키지 구조 변경

* feat: runner 작성

* feat: 기능 구현

* feat: ui구현

* test: 테스트 작성

* chore: 로컬에서는 트레이스 끔

* refactor: dto 필드 이름 수정

* refactor: 명세 수정

* refactor: 명세 수정

* refactor: 패키지 구조 변경

* test: 레이싱 게임 관련 도메인 로직 작성

* refactor: 테스트를 위해 기본속도 10으로 지정

* refactor: 레이싱게임 시작 경로 `/room/{joinCode}/minigame/command` 이용하도록 수정

* feat: 미니게임 시작 및 플로우 기능 작성

* refactor: command로 게임을 시작하면 round에서 실행된 게임의 응답을 반환하도록 수정

* feat: 레이싱게임 스코어 반환하도록 수정

* refactor: 시작점과 결승점에서는 속도가 0이 되도록 수정 / 속도, 위치에 scale가능하도록 수정 / PREPARE전에 플레이어 정보 반환하도록 수정

* refactor: PLAYING으로 바뀐 후 이벤트 쏘도록 수정

* docs: 불필요한 주석 제거

* refactor: 카드게임 태스크 진행 이벤트 기반에서 직접 호출하는 방식으로 수정

* test: racinggame 관련 로직 테스트 작성

* feat: 동점자 계산 수정

* test: 탭 테스트 추가

* feat: 레이싱게임 시작, 종료시 db에 저장

* refactor: 미니게임 저장하는 위치 변경

* test: 레이싱게임 finishTime이 명확히 차이나도록 로직 수정

* refactor: 불필요한 임포트 제거

* chore: 응답 dto 패키지 위치지 변경

* chore: 클로드 md 제거

* chore: lab코드 제거

* refactor: 코드레빗 리뷰 반영

* refactor: 스케줄러 패키지 위치 적절히 변경 / 스레드 이름 프리픽스 수정

* refactor: score를 long으로 변경

* refactor: racingGame service 네이밍 적절히 변경

* refactor: RacingRange 예외처리

* refactor: 시간 외부에서 주입받도록 변경

* refactor: updateSpeed로직 Runner.java로 이동

* refactor: clamp로 수정

* refactor: runnersMovedEvent 정팩메 이름 from → of로 변경

* refactor: 불필요한 createdAt제거

* refactor: 정팩메 컨벤션에 맞게 네이밍 수정

* refactor: 예외 메시지 수정

* refactor: 외부에서 사용하지 않는 메서드 private으로 변경

* refactor: 변경된 정팩메 이름에 맞게 사용처 수정

* refactor: MiniGameResult 정렬 기준에 따라 메서드 분리

* refactor:  MiniGameFinishedEvent에 EventID 추가해서 분산락 지원하도록 수정

---------

+++ backend/src/main/java/coffeeshout/minigame/ui/response/MiniGameScoresResponse.java
@@ -10,11 +10,14 @@ public record MiniGameScoresResponse(List<MiniGameScoreResponse> scores) {
 
     public record MiniGameScoreResponse(
             String playerName,
-            Integer score
+            Long score
     ) {
 
         public static MiniGameScoreResponse from(@NonNull Map.Entry<Player, MiniGameScore> scoreEntry) {
-            return new MiniGameScoreResponse(scoreEntry.getKey().getName().value(), scoreEntry.getValue().getValue());
+            return new MiniGameScoreResponse(
+                    scoreEntry.getKey().getName().value(),
+                    scoreEntry.getValue().getValue()
+            );
         }
     }
